+++ src/components/Form/components/FormEditButton.component.tsx
@@ -11,29 +11,44 @@ import type { FormEditButtonStyleProps } from "../Form.types";
 /**
  * Renders a form's edit button and hides it based on edit state.
  *
- * @version 0.0.2
+ * @version 0.0.3
  * @private
  *
  * @return {JSX.Element} The rendered form edit button component.
  */
 export const FormEditButton = (): JSX.Element => {
-  const { isEditing, startEditing, readonly } = useFormContext();
+  const { isEditing, startEditing, readonly, cancelEdition, asModal } =
+    useFormContext();
   const tl = useUikitTranslation();
 
   const label = useStatic(tl.action.edit);
+  const closeLabel = useStatic(tl.action.close);
 
   return (
-    <EditButton
-      $hidden={isEditing || readonly}
-      onClick={startEditing}
-      icon="Pen"
-      title={label}
-      testId="form-edit-button"
-      overrideTestId
-      label={label}
-      secondary
-      expandOnHover
-    />
+    <>
+      <EditButton
+        $hidden={isEditing || readonly}
+        onClick={startEditing}
+        icon="Pen"
+        title={label}
+        testId="form-edit"
+        label={label}
+        secondary
+        expandOnHover
+      />
+      {asModal && (
+        <EditButton
+          $hidden={!isEditing}
+          onClick={cancelEdition}
+          icon="XLg"
+          title={closeLabel}
+          label={closeLabel}
+          testId="form-modal-close"
+          color="black"
+          expandOnHover
+        />
+      )}
+    </>
   );
 };